docs: rewrite hvsupport.html page generator in python
As part of an goal to eliminate Perl from libvirt build tools,
rewrite the hvsupport.pl tool in Python.

This was a straight conversion, manually going line-by-line to
change the syntax from Perl to Python. Thus the overall structure
of the file and approach is the same.

The new impl generates byte-for-byte identical output to the
old impl.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190911162333.8668-24-berrange@redhat.com>