network: fix connection usage counts after restart
Since the introduction of the virNetworkPort object, the network driver
has a persistent record of ports that have been created against the
networks. Thus the hypervisor drivers no longer communicate to the
network driver during libvirtd restart.

This change, however, meant that the connection usage counts were
no longer re-initialized during a libvirtd restart. To deal with this we
must iterate over all virNetworkPortDefPtr objects we have and invoke
the notify callback to record the connection usage count.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190913145915.12404-1-berrange@redhat.com>
Reviewed-by: Laine Stump <laine@laine.org>