virpci, qemu_domain_address: adding virPCIDeviceSetAddress
A common operation in qemu_domain_address is comparing a
virPCIDeviceAddress and assigning domain, bus, slot and function
to a specific value. The former can be done with the existing
virPCIDeviceAddressEqual() helper.

A new virpci helper called virPCIDeviceSetAddress() was created to
make it easier to assign the same domain/bus/slot/function of an
already existent virPCIDeviceAddress. Using both in
qemu_domain_address made the code tidier, since the object
was already created to use virPCIDeviceAddressEqual().

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20190919200447.7653-3-danielhb413@gmail.com>