qemu: fix operation job association when qemuCaps is NULL.
Function qemuMigrationEatCookie() calls
qemuMigrationCookieXMLParseStr(), when qemuCaps is not set, function
assumes priv as a NULL pointer. At the end, function tries to set
job info operation to the same current job. But, if priv is NULL due to
missing qemuCaps, code returns a Null Pointer Exception. This commit
adds an extra check to verify it.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20190921165522.14574-1-jcfaracco@gmail.com>