build: introduce bare minimum use of meson as a build system
This patch is the first baby step in an incremental conversion
to use meson. Initially all it does is take responsibilty for
creating the dist tarball.

No attempt is made to integrate into autotools at this stage.
Eventually meson will be the primary build system and autotools
will be invoked by meson automatically for building gnulib.
To start with though, the build process will require invoking
both automake and meson. Since meson only supports
builddir != srcdir, automake must be invoked this way too:

   mkdir build
   cd build
   ../configure
   make
   cd ..
   meson build
   ninja -C build

Though in reality the meson/ninja step here is a no-op,
it is enough to start the integration process.

In terms of making a release the steps are the same, but the
last command must be followed by "ninja -C build dist"

The dist tarball we generate is slightly different from what
we previously built. meson's default is to merely bundle the
result of 'git archive' with all submodules. We do not want
the .gnulib submodule though as that is far too large. We
just want the subset of gnulib installed by bootstrap. We
can't simply copy the gnulib/ directory though, as we need
the autotools integration too and bootstrap runs autoreconf.

The upshot is that we purge .gnulib and copy a whitelist of
files that were seen to be created by bootstrap in a Fedora
30 host. By the end of the series, when gnulib is the only
thing using autotools, we can isolate everything from
bootstrap in a private sub-dir and simplify the dist.

In addition to the gnulib / autotools pieces, we add the
AUTHORS, NEWS and libvirt.spec files.

What this looses compared to previous tarballs is any
files that were listed as EXTRA_DIST, which were not
source controlled in git. This is negligible and any
such files are easily built.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190927111158.20653-4-berrange@redhat.com>