tools: make virt-host-validate check CPU vulnerabilities
Add a check reporting if any CPU vulnerabilities have not been mitigated
by the kernel. It further reports whether it is safe to use Intel SMT
for KVM guests or not, as several of the vulnerabilities are dangerous
when combined with SMT and KVM, even if mitigations are in effect.

eg on a host with mitigations, but unsafe SMT still enabled:

  Checking CPU hardware vulnerability mitigation...PASS
  Checking CPU hardware vulnerability SMT safety...FAIL

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20190927125225.22432-7-berrange@redhat.com>